V4: Fix array-like length truncation in Array/TypedArray builtins
Object::getLength() returns the abstract length of an array-like as a
qint64: for a real Array it is bounded by the exotic length invariant
(<= 2^32-1), but for a generic object accessed via
Array.prototype.<method>.call(obj, ...) it is the result of ToLength()
and can legitimately be as large as 2^53-1 (e.g. a plain object with a
custom "length" property).

Most Array.prototype methods immediately truncated that value into a
uint, so a length at or beyond 2^32 silently wrapped
- e.g. length: 2**32, 0: 12}` truncated to a length of 0, making
Array.prototype.some.call() wrongly return false instead of true.

Widen len, and the loop counters compared against it, to qint64
throughout both files so the value survives intact. Also add the missing
RangeError check in Array.prototype.slice for when the result would need
more than 2^32-1 elements, matching the ArrayCreate step of the spec
algorithm that was missing entirely.

Pick-to: 6.12

+++ src/qml/jsruntime/qv4typedarray.cpp
@@ -1753,7 +1753,7 @@ ReturnedValue IntrinsicTypedArrayCtor::method_from(const FunctionObject *f, cons
         if (!arrayLike)
             return scope.engine->throwTypeError(QString::fromLatin1("Cannot convert %1 to object").arg(argv[0].toQStringNoThrow()));
 
-        int len = arrayLike->getLength();
+        qint64 len = arrayLike->getLength();
         CHECK_EXCEPTION();
 
         // Getting the length may throw, and must do so before we check the constructor validity.
@@ -1761,7 +1761,7 @@ ReturnedValue IntrinsicTypedArrayCtor::method_from(const FunctionObject *f, cons
             return scope.engine->throwTypeError();
 
         ScopedObject a(scope, Value::undefinedValue());
-        ScopedValue ctorArgument(scope, Value::fromReturnedValue(QV4::Encode(len)));
+        ScopedValue ctorArgument(scope, Value::fromReturnedValue(QV4::Encode(double(len))));
         a = C->callAsConstructor(ctorArgument, 1);
         CHECK_EXCEPTION();
 
@@ -1771,7 +1771,7 @@ ReturnedValue IntrinsicTypedArrayCtor::method_from(const FunctionObject *f, cons
 
         ScopedValue mappedValue(scope, Value::undefinedValue());
         ScopedValue kValue(scope);
-        for (int k = 0; k < len; ++k) {
+        for (qint64 k = 0; k < len; ++k) {
             kValue = arrayLike->get(k);
             CHECK_EXCEPTION();
